# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: stocks
    description: Stock quant data operations
  - name: screeners
    description: Thematic Top 10 stock screener operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /stocks/search:
    get:
      operationId: searchStocks
      tags: [stocks]
      summary: Search for stock tickers
      parameters:
        - name: q
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Matching tickers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/StockSearchResult"
  /stocks/popular:
    get:
      operationId: getPopularStocks
      tags: [stocks]
      summary: List a curated set of popular tickers
      responses:
        "200":
          description: Popular tickers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/StockSearchResult"
  /stocks/{ticker}:
    get:
      operationId: getStockDetail
      tags: [stocks]
      summary: Get full quant detail for a ticker
      parameters:
        - name: ticker
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Stock detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StockDetail"
        "404":
          description: Ticker not found
  /stocks/{ticker}/historical-valuation:
    get:
      operationId: getHistoricalValuation
      tags: [stocks]
      summary: Get historical PE/PB valuation bands for a ticker
      parameters:
        - name: ticker
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Historical valuation bands
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HistoricalValuation"
        "404":
          description: Ticker not found
  /stocks/{ticker}/ai-committee:
    get:
      operationId: getAiCommitteeVerdict
      tags: [stocks]
      summary: Get the AI Investment Committee verdict for a ticker
      parameters:
        - name: ticker
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: AI committee verdict
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AiCommitteeVerdict"
        "404":
          description: Ticker not found
  /stocks/{ticker}/fund-flow:
    get:
      operationId: getFundFlow
      tags: [stocks]
      summary: Get institutional ownership and volume shock data for a ticker
      parameters:
        - name: ticker
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Fund and whale flow data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FundFlow"
        "404":
          description: Ticker not found
  /screeners:
    get:
      operationId: getScreeners
      tags: [screeners]
      summary: Get the 5 thematic Top 10 stock screener leaderboards
      responses:
        "200":
          description: Screener leaderboards
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ScreenerList"
components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status
    StockSearchResult:
      type: object
      required: [symbol, name]
      properties:
        symbol:
          type: string
        name:
          type: string
        exchange:
          type: ["string", "null"]
    MarketData:
      type: object
      required:
        - price
        - change
        - changePercent
        - marketCap
        - peRatio
        - forwardPe
        - pegRatio
        - epsTtm
        - roe
        - freeCashFlow
        - grossMargin
        - currency
        - lastUpdated
      properties:
        price: { type: ["number", "null"] }
        change: { type: ["number", "null"] }
        changePercent: { type: ["number", "null"] }
        marketCap: { type: ["number", "null"] }
        peRatio: { type: ["number", "null"] }
        forwardPe: { type: ["number", "null"] }
        pegRatio: { type: ["number", "null"] }
        epsTtm: { type: ["number", "null"] }
        roe: { type: ["number", "null"] }
        freeCashFlow: { type: ["number", "null"] }
        grossMargin: { type: ["number", "null"] }
        currency: { type: string }
        lastUpdated: { type: string, format: date-time }
    GuruScores:
      type: object
      required:
        - buffettScore
        - buffettSummary
        - lynchScore
        - lynchSummary
        - mungerScore
        - mungerSummary
      properties:
        buffettScore: { type: number }
        buffettSummary: { type: string }
        lynchScore: { type: number }
        lynchSummary: { type: string }
        mungerScore: { type: number }
        mungerSummary: { type: string }
    DcfValuation:
      type: object
      required:
        - fairValue
        - currentPrice
        - marginOfSafetyPercent
        - expectedReturn1Y
        - expectedReturn3Y
        - expectedReturn5Y
        - assumedGrowthRate
        - discountRate
        - terminalGrowthRate
      properties:
        fairValue: { type: ["number", "null"] }
        currentPrice: { type: ["number", "null"] }
        marginOfSafetyPercent: { type: ["number", "null"] }
        expectedReturn1Y: { type: ["number", "null"] }
        expectedReturn3Y: { type: ["number", "null"] }
        expectedReturn5Y: { type: ["number", "null"] }
        assumedGrowthRate: { type: ["number", "null"] }
        discountRate: { type: number }
        terminalGrowthRate: { type: number }
    TradingSignals:
      type: object
      required:
        - suggestedBuyPrice
        - stopLoss
        - targetPrice1
        - targetPrice2
        - rsi14
        - sma50
        - sma200
        - trend
      properties:
        suggestedBuyPrice: { type: ["number", "null"] }
        stopLoss: { type: ["number", "null"] }
        targetPrice1: { type: ["number", "null"] }
        targetPrice2: { type: ["number", "null"] }
        rsi14: { type: ["number", "null"] }
        sma50: { type: ["number", "null"] }
        sma200: { type: ["number", "null"] }
        trend:
          type: string
          enum: [bullish, bearish, neutral]
    NewsItem:
      type: object
      required: [title, link, publisher, publishedAt]
      properties:
        title: { type: string }
        link: { type: string }
        publisher: { type: string }
        publishedAt: { type: string, format: date-time }
    FinancialLinks:
      type: object
      required: [incomeStatementUrl, balanceSheetUrl, cashFlowUrl, secFilingsUrl]
      properties:
        incomeStatementUrl: { type: string }
        balanceSheetUrl: { type: string }
        cashFlowUrl: { type: string }
        secFilingsUrl: { type: string }
    ExperimentalScores:
      type: object
      required:
        - aiBenefitScore
        - aiBenefitSummary
        - sorosScore
        - sorosSummary
        - dalioScore
        - dalioSummary
        - eps2027Estimate
      properties:
        aiBenefitScore: { type: number }
        aiBenefitSummary: { type: string }
        sorosScore: { type: number }
        sorosSummary: { type: string }
        dalioScore: { type: number }
        dalioSummary: { type: string }
        eps2027Estimate: { type: ["number", "null"] }
    StockDetail:
      type: object
      required:
        - ticker
        - companyName
        - sector
        - industry
        - marketData
        - guruScores
        - dcfValuation
        - tradingSignals
        - news
        - financialLinks
        - experimentalScores
      properties:
        ticker: { type: string }
        companyName: { type: string }
        sector: { type: ["string", "null"] }
        industry: { type: ["string", "null"] }
        marketData: { $ref: "#/components/schemas/MarketData" }
        guruScores: { $ref: "#/components/schemas/GuruScores" }
        dcfValuation: { $ref: "#/components/schemas/DcfValuation" }
        tradingSignals: { $ref: "#/components/schemas/TradingSignals" }
        news:
          type: array
          items:
            $ref: "#/components/schemas/NewsItem"
        financialLinks: { $ref: "#/components/schemas/FinancialLinks" }
        experimentalScores: { $ref: "#/components/schemas/ExperimentalScores" }
    ValuationBandLevels:
      type: object
      required: [minus2SD, minus1SD, mean, plus1SD, plus2SD]
      properties:
        minus2SD: { type: ["number", "null"] }
        minus1SD: { type: ["number", "null"] }
        mean: { type: ["number", "null"] }
        plus1SD: { type: ["number", "null"] }
        plus2SD: { type: ["number", "null"] }
    ValuationPeriodStats:
      type: object
      required:
        - years
        - currentPe
        - pePercentile
        - peMean
        - peStdDev
        - priceBandsFromPe
        - currentPb
        - pbPercentile
        - pbMean
        - pbStdDev
        - priceBandsFromPb
      properties:
        years: { type: number }
        currentPe: { type: ["number", "null"] }
        pePercentile: { type: ["number", "null"] }
        peMean: { type: ["number", "null"] }
        peStdDev: { type: ["number", "null"] }
        priceBandsFromPe: { $ref: "#/components/schemas/ValuationBandLevels" }
        currentPb: { type: ["number", "null"] }
        pbPercentile: { type: ["number", "null"] }
        pbMean: { type: ["number", "null"] }
        pbStdDev: { type: ["number", "null"] }
        priceBandsFromPb: { $ref: "#/components/schemas/ValuationBandLevels" }
    ValuationChartPoint:
      type: object
      required:
        - date
        - price
        - peBandMinus2SD
        - peBandMinus1SD
        - peBandMean
        - peBandPlus1SD
        - peBandPlus2SD
      properties:
        date: { type: string }
        price: { type: number }
        peBandMinus2SD: { type: ["number", "null"] }
        peBandMinus1SD: { type: ["number", "null"] }
        peBandMean: { type: ["number", "null"] }
        peBandPlus1SD: { type: ["number", "null"] }
        peBandPlus2SD: { type: ["number", "null"] }
    HistoricalValuation:
      type: object
      required:
        - ticker
        - isHeuristic
        - methodologyNote
        - currentPrice
        - currentPe
        - currentPb
        - periods
        - chartSeries
      properties:
        ticker: { type: string }
        isHeuristic: { type: boolean }
        methodologyNote: { type: string }
        currentPrice: { type: ["number", "null"] }
        currentPe: { type: ["number", "null"] }
        currentPb: { type: ["number", "null"] }
        periods:
          type: object
          required: [threeYear, fiveYear, tenYear]
          properties:
            threeYear:
              oneOf:
                - $ref: "#/components/schemas/ValuationPeriodStats"
                - type: "null"
            fiveYear:
              oneOf:
                - $ref: "#/components/schemas/ValuationPeriodStats"
                - type: "null"
            tenYear:
              oneOf:
                - $ref: "#/components/schemas/ValuationPeriodStats"
                - type: "null"
        chartSeries:
          type: array
          items:
            $ref: "#/components/schemas/ValuationChartPoint"
    AiCommitteeVerdict:
      type: object
      required:
        - ticker
        - isHeuristic
        - methodologyNote
        - agentVerdict
        - confidence
        - reasoningBreakdown
        - keyCatalysts
        - hiddenRisks
      properties:
        ticker: { type: string }
        isHeuristic: { type: boolean }
        methodologyNote: { type: string }
        agentVerdict:
          type: string
          enum: [BUY, HOLD, SELL]
        confidence: { type: number }
        reasoningBreakdown:
          type: object
          required:
            - valueView
            - valueStance
            - growthView
            - growthStance
            - macroView
            - macroStance
          properties:
            valueView: { type: string }
            valueStance:
              type: string
              enum: [bullish, bearish, neutral]
            growthView: { type: string }
            growthStance:
              type: string
              enum: [bullish, bearish, neutral]
            macroView: { type: string }
            macroStance:
              type: string
              enum: [bullish, bearish, neutral]
        keyCatalysts:
          type: array
          items: { type: string }
        hiddenRisks:
          type: array
          items: { type: string }
    WhaleHolder:
      type: object
      required: [organization, pctHeld, shares, value, pctChange]
      properties:
        organization: { type: string }
        pctHeld: { type: ["number", "null"] }
        shares: { type: ["number", "null"] }
        value: { type: ["number", "null"] }
        pctChange: { type: ["number", "null"] }
    FundFlow:
      type: object
      required:
        - ticker
        - isHeuristic
        - methodologyNote
        - institutionsPercentHeld
        - institutionsCount
        - insidersPercentHeld
        - institutionalMomentum
        - institutionalMomentumNote
        - topWhaleHolders
        - currentVolume
        - averageVolume
        - volumeRatio
        - priceChangePercent
        - volumeShockStatus
        - volumeShockNote
      properties:
        ticker: { type: string }
        isHeuristic: { type: boolean }
        methodologyNote: { type: string }
        institutionsPercentHeld: { type: ["number", "null"] }
        institutionsCount: { type: ["number", "null"] }
        insidersPercentHeld: { type: ["number", "null"] }
        institutionalMomentum:
          type: string
          enum: [increasing, decreasing, flat, unknown]
        institutionalMomentumNote: { type: string }
        topWhaleHolders:
          type: array
          items:
            $ref: "#/components/schemas/WhaleHolder"
        currentVolume: { type: ["number", "null"] }
        averageVolume: { type: ["number", "null"] }
        volumeRatio: { type: ["number", "null"] }
        priceChangePercent: { type: ["number", "null"] }
        volumeShockStatus:
          type: string
          enum: [accumulation, distribution, normal]
        volumeShockNote: { type: string }
    ScreenerStock:
      type: object
      required: [symbol, name, price, metricLabel, metricValue, rank]
      properties:
        symbol: { type: string }
        name: { type: string }
        price: { type: ["number", "null"] }
        metricLabel: { type: string }
        metricValue: { type: ["number", "null"] }
        rank: { type: integer }
    ScreenerList:
      type: object
      required: [key, title, description, stocks]
      properties:
        key: { type: string }
        title: { type: string }
        description: { type: string }
        stocks:
          type: array
          items:
            $ref: "#/components/schemas/ScreenerStock"