# lib/api-zod/src/generated/types/aiCommitteeVerdict.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { AiCommitteeVerdictAgentVerdict } from './aiCommitteeVerdictAgentVerdict';
import type { AiCommitteeVerdictReasoningBreakdown } from './aiCommitteeVerdictReasoningBreakdown';

export interface AiCommitteeVerdict {
  ticker: string;
  isHeuristic: boolean;
  methodologyNote: string;
  agentVerdict: AiCommitteeVerdictAgentVerdict;
  confidence: number;
  reasoningBreakdown: AiCommitteeVerdictReasoningBreakdown;
  keyCatalysts: string[];
  hiddenRisks: string[];
}