# lib/api-zod/src/generated/types/fundFlow.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { FundFlowInstitutionalMomentum } from './fundFlowInstitutionalMomentum';
import type { FundFlowVolumeShockStatus } from './fundFlowVolumeShockStatus';
import type { WhaleHolder } from './whaleHolder';

export interface FundFlow {
  ticker: string;
  isHeuristic: boolean;
  methodologyNote: string;
  /** @nullable */
  institutionsPercentHeld: number | null;
  /** @nullable */
  institutionsCount: number | null;
  /** @nullable */
  insidersPercentHeld: number | null;
  institutionalMomentum: FundFlowInstitutionalMomentum;
  institutionalMomentumNote: string;
  topWhaleHolders: WhaleHolder[];
  /** @nullable */
  currentVolume: number | null;
  /** @nullable */
  averageVolume: number | null;
  /** @nullable */
  volumeRatio: number | null;
  /** @nullable */
  priceChangePercent: number | null;
  volumeShockStatus: FundFlowVolumeShockStatus;
  volumeShockNote: string;
}