# lib/api-zod/src/generated/types/historicalValuation.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { HistoricalValuationPeriods } from './historicalValuationPeriods';
import type { ValuationChartPoint } from './valuationChartPoint';

export interface HistoricalValuation {
  ticker: string;
  isHeuristic: boolean;
  methodologyNote: string;
  /** @nullable */
  currentPrice: number | null;
  /** @nullable */
  currentPe: number | null;
  /** @nullable */
  currentPb: number | null;
  periods: HistoricalValuationPeriods;
  chartSeries: ValuationChartPoint[];
}