# lib/api-zod/src/generated/types/valuationBandLevels.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ValuationBandLevels {
  /** @nullable */
  minus2SD: number | null;
  /** @nullable */
  minus1SD: number | null;
  /** @nullable */
  mean: number | null;
  /** @nullable */
  plus1SD: number | null;
  /** @nullable */
  plus2SD: number | null;
}