# lib/api-zod/src/generated/types/valuationChartPoint.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ValuationChartPoint {
  date: string;
  price: number;
  /** @nullable */
  peBandMinus2SD: number | null;
  /** @nullable */
  peBandMinus1SD: number | null;
  /** @nullable */
  peBandMean: number | null;
  /** @nullable */
  peBandPlus1SD: number | null;
  /** @nullable */
  peBandPlus2SD: number | null;
}