# lib/api-zod/src/generated/types/valuationPeriodStats.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ValuationBandLevels } from './valuationBandLevels';

export interface ValuationPeriodStats {
  years: number;
  /** @nullable */
  currentPe: number | null;
  /** @nullable */
  pePercentile: number | null;
  /** @nullable */
  peMean: number | null;
  /** @nullable */
  peStdDev: number | null;
  priceBandsFromPe: ValuationBandLevels;
  /** @nullable */
  currentPb: number | null;
  /** @nullable */
  pbPercentile: number | null;
  /** @nullable */
  pbMean: number | null;
  /** @nullable */
  pbStdDev: number | null;
  priceBandsFromPb: ValuationBandLevels;
}