# lib/api-zod/src/generated/types/whaleHolder.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface WhaleHolder {
  organization: string;
  /** @nullable */
  pctHeld: number | null;
  /** @nullable */
  shares: number | null;
  /** @nullable */
  value: number | null;
  /** @nullable */
  pctChange: number | null;
}